orientation-event/DeviceMotion/t031.html

<!DOCTYPE html>
<html>

  <head>
	<title>Implementations that are unable to provide the accelerationIncludingGravity property must set the values to null:</title>
	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	<script src="http://w3c-test.org/resources/testharness.js"></script>
	<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
	<script src="http://w3c-test.org/resources/WebIDLParser.js"></script>
	<script src="http://w3c-test.org/resources/idlharness.js"></script>
  <link rel='author' title='Mosquito FP7'/>
  </head>

  <body>

  
  	<h1>Implementations that are unable to provide the accelerationIncludingGravity property must set the values to null:</h1>
	<div id="log">
	</div>
	
		<script type="text/javascript">
		tgvt = async_test("accelerationIncludingGravity property must be set to null");
		tx = async_test("x property of accelerationIncludingGravity must be set to null");
		ty = async_test("y property of accelerationIncludingGravity must be set to null");
		tz = async_test("z property of accelerationIncludingGravity must be set to null");

		vgvt = "";
		vx = "";
		vy = "";
		vz = "";
		window.addEventListener('devicemotion', function (e){
			acc = e.accelerationIncludingGravity;
			vgvt = acc;
			vx = acc.x;
			vy = acc.y;
			vz = acc.z;
		}, false);
		
		setTimeout("testing()", 200);
		
		function testing(){
			tgvt.step(function() {assert_equals(vgvt, null)});
			tgvt.done();
			tx.step(function() {assert_equals(vx, null)});
			tx.done();
			ty.step(function() {assert_equals(vy, null)});
			ty.done();
			tz.step(function() {assert_equals(vz, null)});				
			tz.done();
		}
	</script>
	
  </body>
  
</html>
